Run testSerializeViaString on all tests
…including the corner cases.

This uncovers ceylon/ceylon-sdk#655, so we assume the bug doesn’t occur
so that the tests still pass. (You’ll notice that this assumption checks
the value of the argument, and so the test is only skipped for that one
particular argument, whereas the assumption above it, for
ceylon/ceylon#6908, triggers on any argument and thus completely
disables that test on JS. I think looking at the value is more
acceptable on an Integer than on a Float, and in any case, I just don’t
want to completely disable that test on the JVM, period. On the other
hand, JS tests are currently broken anyways due to ceylon/ceylon#6984.)