Allow deserializing unions of JSON primitive types
In Ceylon, Integer, String etc. are classes, so the previous version of
the assertion didn’t allow e. g. `Integer|String`.

To avoid copying the stream of JSON primitive types another time, it is
extracted into a new variable (which is even shared, because we’re nice
like that). The variable is also a Set instead of just a stream, which
significantly speeds up a membership test: according to a quick’n’dirty
benchmark, about sixfold for `Entry<Integer,String>` (not in the set),
tenfold for `Null` (last element of the stream), and a whopping
fortyfold for `Integer` (first element of the stream). I’m not sure if
those results make sense (`Integer` in stream is slightly faster than
`Null` in stream, but `Integer` in set is a lot *more* faster than
`Null` in set – why?), but in any event, it’s a performance boost.